+++ src/TestTrackingDiagrams/Tracking/RequestResponseLog.cs
@@ -17,7 +17,8 @@ public record RequestResponseLog(
     bool TrackingIgnore,
     OneOf<HttpStatusCode, string>? StatusCode = null,
     RequestResponseMetaType MetaType = default,
-    string? DependencyCategory = null)
+    string? DependencyCategory = null,
+    string? CallerDependencyCategory = null)
 {
     public bool IsOverrideStart { get; set; }
     public bool IsOverrideEnd { get; set; }